Add files via upload
Hey guys,
This is a case study which contains questions based on MySQL joins. I created them in the format of question and answers. Raw data is in the csv files. I imported the data using MySQL table import wizard . Cleaned the data and then later worked on the questions given in the cases study.

Regards